pyproject.toml
[tool.poetry]
name = "Medical_Appointment_Scheduling_API"
version = "0.1.0"
description = "API Django para gestão de consultas medicas"
authors = ["Kaio Herculano <kaioherculano12@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "3.9.1"
Django = "5.2.4"
djangorestframework = "3.16.0"
djangorestframework-simplejwt = "5.5.0"
PyJWT = "2.9.0"
sqlparse = "0.5.3"
tzdata = "2025.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
gunicorn = "^20.1.0"
dj-database-url = "^1.0.0"
whitenoise = "^6.9.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"
pytest = "^7.0"
black = "^23.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
